Update README.txt
The call back url is not documented anywhere. I talked to a coworker doing wp themes and said they dropped this plugin because they couldn't figure out where they supposed to use as the callback. I went on searching for it and after a few tries I got the right url. It doesn't hurt to put it out there to make it easier for anyone struggling with the same.